requirements.txt
# === Web Framework ===
fastapi==0.110.0
uvicorn[standard]==0.29.0

# === WebSocket 客户端 ===
websockets==12.0

# === 异步任务处理 ===
celery==5.3.6
redis==5.0.3

# === 环境变量加载 ===
python-dotenv==1.0.1

# === 数据模型和验证 ===
pydantic==2.11.7
pydantic-settings==2.10.1

# === HTTP 请求（用于 AI API 调用） ===
httpx==0.27.0

# === 数据库操作（可选）===
psycopg2-binary==2.9.9  # PostgreSQL 连接驱动

# === 时间处理 ===
pendulum==3.0.0